Some additional changes to test and error response
